# Add test and benchmark for `FindOffset`