Low-hanging fruit for improving test coverage

```
benchmark                                                         old ns/op     new ns/op     delta
BenchmarkFindOffset/FindOffset/Null_Inputs-8                      9.55          3.33          -65.13%
BenchmarkFindOffset/FindOffset/Longer_Inputs-8                    2052          1823          -11.16%
BenchmarkFindOffset/FindOffset/Early_but_beyond_end_of_line-8     19450         168           -99.14%
```